committed path by K2, mail by him follows:
I was messing with stage this afternoon and noticed it seg faulting after
its call to GetCodec. When called, GetCodec simply dereferences the top of
the outCodecs list; if the list is empty, however, a seg fault occours.
This situation was arising due to the fact that during polling the server
was failing to find any codecs in common with the client and yet the
GetStatus method would still return successful.

I've added a test for empty listedness to GetStatus; It is prolly prudent
to add one to GetCodec as well (just in case). GetCodec should prolly
throw an exception if outCodecs is empty, but all the exception stuff is
isolated in amoungst the Object stuff so I made it return NULL for the
time being - something to think about tho.

+ additional patch by k2:

[17:10:49] <k2> if the client disconnects during polling the socket enters
an illigal state - all further socket.gets() return imediately....
[17:11:44] <k2> the result is the connection never enters a done state but
never fails either