Delegate allocation to separate class.
This aims to make the boilerplate for each BaseObject subclass instance
smaller and concentrated at templated code, so that further improvements
are possible.

The main change is that each subclass of BaseObject now needs to keep
onto a static field named "allocator" which is of the type
Allocator<subclass>. They also need to implement a static method named
fillDefaultObjectInstance which is responsible for creating the
prototype instance and set the prototype attribute flags.