Added an element exerciser class for testing.
This is to be used when testing any method which parses Atlas data sent
over the network. Since any such method cannot crash or throw an
exception if the data is incorrect (remember kids: never trust the
network) we should be able to send any kind of bogus data to it with
success.