Move initialization back to implementation.
This is because we need to support MSVC which currently doesn't support
"constexpr".
Note that the reason why we can't initialize these values in the header
is that C++ only allows that to happen for integral types, which "float"
is not.
Whenever MSVC gains the ability to use "constexpr" this commit should be
reverted.