Added a test for terrain generation.
It's important that the algorithm used by Mercator to generate terrain
is unchanged between releases. This test performs a simple check to make
sure that the algorithm still is valid. We do this by sampling three
points. Given the way the terrain is generated we can be pretty sure
that if something related to the algorithm is changed at least one of
these three values will be different.