Properly implement test.
Bach isn't correctly escaping characters, so that's disabled for now.
However, "packed" and "xml" are important to test, since these are most
used.
"packed" is used for all network traffic, and "xml" is used for
persistence on the server and client. "bach" however isn't actually used
currently as far as I know.