Handle the avatar entity being deleted.
When this happens we should deactivate the character, since it relies on
the entity being available in order to experience the world.
We need to devise a future system where an entity can be defeated
without being destroyed; this to handle the typical scenario in a
virtual world where the client isn't immediately disconnected when the
avatar entity is killed.