Use the io_service for timed event.
We've done away with our own implementation of TimedEventService and
Timeout and instead use boost::asio::io_service.
The semantics of how the library is used is also slightly changed. You
must now create an instance of TimedEventService yourself at startup. We
probably want to change this so it's not a singleton; but for now this
will have to do.