Handle entities appearing by Move op only.
This signals a change in the protocol. Previously we had the server
first send an Appear op for any entity that moved within our field of
view. However, we now only send a Move op, and let the client handle it
as an appearance.
I.e. from the clients point of view, if it receives a Move operation
from an entity that's either not known, or known but not visible, it
means that this entity has moved into the field of view.