Remove the Factories static instance.
When an instance of Factories in needed, such as when setting the "args"
attribute, one must be supplied. Normally this only occurs in the
decoder stage, so for normal calls to "setAttr" a null value can be
supplied.
This removes the last static singleton from the Atlas library, allowing
it to work with multiple independent factory instances.