Make Generator skip existing files.
This is done by first asking if for a map of existing files, and then comparing
that to the source directory. When doing this we compare the write times of the files.
This means that we need to store the data in our repository with the same write
times as the source data.