Sinmplify op handling.
Instead of trying to figure out how to handle incoming ops for entities we
now just keep them in a queue until we've resolved the entity and replays them.
This helps with the client since it means that we can be sure to handle ops once
the client has had time to figure out which Model to use.