Allow widgets to be built as reloadable plugins.
This is useful when doing development. It requires some libs to be built as shared though.
Enable this by specifying the "widgets_as_plugins" when invoking conan.
The reason we need to do this in the Conan phase is because it affects how some
libraries now needs to be built as shared libs.